Move the DockerExecutor to tests
