+++ docker-compose.yml
@@ -46,7 +46,7 @@ services:
     environment:
       <<: [*postgres-env, *minio-env]
       COMPONENTS_REGISTRY_INSECURE: 'true'
-      COMPONENTS_DEFAULT_BACKEND: 'grandchallenge.components.backends.docker.DockerExecutor'
+      COMPONENTS_DEFAULT_BACKEND: 'tests.components_tests.resources.backends.InsecureDockerExecutor'
       COMPONENTS_STRIP_LEADING_PREFIX_SLASH: 'true'
       DEBUG: 'true'
       COMPRESS_OFFLINE: 'false'
@@ -108,7 +108,7 @@ services:
     environment:
       <<: [*postgres-env, *minio-env]
       COMPONENTS_REGISTRY_INSECURE: 'true'
-      COMPONENTS_DEFAULT_BACKEND: 'grandchallenge.components.backends.docker.DockerExecutor'
+      COMPONENTS_DEFAULT_BACKEND: 'tests.components_tests.resources.backends.InsecureDockerExecutor'
       COMPONENTS_STRIP_LEADING_PREFIX_SLASH: 'true'
       DEBUG: 'true'
       PYTHONDONTWRITEBYTECODE: 1
@@ -137,7 +137,7 @@ services:
       <<: [*postgres-env, *minio-env]
       REMAP_SIGTERM: 'SIGQUIT'
       COMPONENTS_REGISTRY_INSECURE: 'true'
-      COMPONENTS_DEFAULT_BACKEND: 'grandchallenge.components.backends.docker.DockerExecutor'
+      COMPONENTS_DEFAULT_BACKEND: 'tests.components_tests.resources.backends.InsecureDockerExecutor'
       COMPONENTS_STRIP_LEADING_PREFIX_SLASH: 'true'
       DEBUG: 'true'
       PYTHONDONTWRITEBYTECODE: 1
